Experiment: call ZigString__freeGlobal when calling deref on an external string
