+++ src/bun.js/bindings/helpers.h
@@ -26,42 +26,6 @@ extern "C" const char* Bun__errnoName(int);
 
 namespace Zig {
 
-// 8 bit byte
-// we tag the final two bits
-// so 56 bits are copied over
-// rest we zero out for consistentcy
-static const unsigned char* untag(const unsigned char* ptr)
-{
-    return reinterpret_cast<const unsigned char*>(
-        (((reinterpret_cast<uintptr_t>(ptr) & ~(static_cast<uint64_t>(1) << 63) & ~(static_cast<uint64_t>(1) << 62)) & ~(static_cast<uint64_t>(1) << 61)) & ~(static_cast<uint64_t>(1) << 60)));
-}
-
-static void* untagVoid(const unsigned char* ptr)
-{
-    return const_cast<void*>(reinterpret_cast<const void*>(untag(ptr)));
-}
-
-static void* untagVoid(const char16_t* ptr)
-{
-    return untagVoid(reinterpret_cast<const unsigned char*>(ptr));
-}
-
-static bool isTaggedUTF16Ptr(const unsigned char* ptr)
-{
-    return (reinterpret_cast<uintptr_t>(ptr) & (static_cast<uint64_t>(1) << 63)) != 0;
-}
-
-// Do we need to convert the string from UTF-8 to UTF-16?
-static bool isTaggedUTF8Ptr(const unsigned char* ptr)
-{
-    return (reinterpret_cast<uintptr_t>(ptr) & (static_cast<uint64_t>(1) << 61)) != 0;
-}
-
-static bool isTaggedExternalPtr(const unsigned char* ptr)
-{
-    return (reinterpret_cast<uintptr_t>(ptr) & (static_cast<uint64_t>(1) << 62)) != 0;
-}
-
 static void free_global_string(void* str, void* ptr, unsigned len)
 {
     // i don't understand why this happens
